Updated validations for multi select option.
Decoupled minLength and maxLength from required as there
can be cases where the item selection is not required but
there has to be a limit on maximum number of elements that
can be selected. So, if maxLength or minLength is defined
then it will be enforced irrespective of whether input is
required or not.